+++ src/__tests__/window.test.ts
@@ -155,7 +155,7 @@ describe("window", () => {
       expect(authentication).toEqual("test-auth-code");
     });
 
-    it("fails if auth window is closed", async () => {
+    it("returns error object if window is closed", async () => {
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
       const childWindow = new Window() as any;
 
@@ -170,13 +170,12 @@ describe("window", () => {
         value: true,
       });
 
-      expect(
-        async () =>
-          await handleChildWindow({
-            url: new URL("https://auth.sbx.laterpaytest.net"),
-            childWindow,
-          }),
-      ).toThrow("window closed");
+      handleChildWindow({
+        url: new URL("https://auth.sbx.laterpaytest.net"),
+        childWindow,
+      }).then((res) => {
+        expect(res).toEqual({ error: "Window closed" });
+      });
     });
   });
 });