+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# [2.0.0](https://github.com/laterpay/supertab-browser/compare/v1.13.0...v2.0.0) (2024-01-29)
+
+### Features
+
+- resolve promise on window close ([#42](https://github.com/laterpay/supertab-browser/issues/42)) ([594dc23](https://github.com/laterpay/supertab-browser/commit/594dc23502e27a1818ace803ab43dd043a96398d))
+
+### BREAKING CHANGES
+
+- Closed windows no longer throw errors but resolve the promise with `{ error: "Window closed" }` instead.
+
+# [1.13.0](https://github.com/laterpay/supertab-browser/compare/v1.12.1...v1.13.0) (2024-01-25)
+
+### Features
+
+- add build for stg tapper environment ([#43](https://github.com/laterpay/supertab-browser/issues/43)) ([57d6d7b](https://github.com/laterpay/supertab-browser/commit/57d6d7befc36cc7bd82d809d94ac0385c5a3f850))
+
 ## [1.12.1](https://github.com/laterpay/supertab-browser/compare/v1.12.0...v1.12.1) (2024-01-19)
 
 ### Bug Fixes